Update tests after breaking changes in master
