+++ src/course-home/data/__factories__/index.js
@@ -1,0 +1,3 @@
+import './courseHomeMetadata.factory';
+import './datesTabData.factory';
+import './outlineTabData.factory';
