+++ src/courseware/data/__factories__/courseBlocks.factory.js
