+++ src/courseware/data/__factories__/courseMetadata.factory.js
