+++ src/courseware/data/__factories__/index.js
@@ -1,7 +1,4 @@
 import './block.factory';
 import './courseBlocks.factory';
-import './courseHomeMetadata.factory';
 import './courseMetadata.factory';
-import './datesTabData.factory';
-import './outlineTabData.factory';
 import './sequenceMetadata.factory';
