+++ src/courseware/data/__factories__/sequenceMetadata.factory.js
